chore: refactor out handler to install page
