+++ src/components/AppInfoConfigOption.js
@@ -4,26 +4,18 @@ import { useQueryParams } from '@newrelic/gatsby-theme-newrelic';
 import MDXContainer from './MDXContainer';
 import TileSelect from './TileSelect';
 
-const AppInfoConfigOption = ({ optionType, selectOptions, mdx }) => {
-  const { queryParams, setQueryParam, deleteQueryParam } = useQueryParams();
+const AppInfoConfigOption = ({ optionType, selectOptions, mdx, onChange }) => {
+  const { queryParams } = useQueryParams();
   const select = selectOptions.find(
     (select) => select.optionType === optionType
   );
-
-  const handleChange = (value) => {
-    if (value !== null || value !== undefined) {
-      setQueryParam(select.optionType, value);
-    } else {
-      deleteQueryParam(select.optionType, value);
-    }
-  };
 
   const { body } = mdx;
   return (
     <div>
       <TileSelect
         label={select.label}
-        onChange={handleChange}
+        onChange={onChange}
         value={
           queryParams.has(select.optionType)
             ? queryParams.get(select.optionType)
@@ -52,6 +44,7 @@ AppInfoConfigOption.propTypes = {
   ).isRequired,
   mdx: PropTypes.object,
   optionType: PropTypes.string.isRequired,
+  onChange: PropTypes.func.isRequired,
 };
 
 export default AppInfoConfigOption;
